+++ packages/spor-alert-react/package.json
@@ -18,10 +18,13 @@
   "scripts": {
     "build": "tsup src/* --dts --treeshake --format esm,cjs",
     "dev": "tsup src/index.tsx --watch --format esm",
-    "test": "vitest run --config ../config/vitest.config.ts"
+    "test": "vitest run --config ../config/vitest.config.ts",
+    "test:watch": "vitest --config ../config/vitest.config.ts"
   },
   "dependencies": {
     "@vygruppen/spor-accordion-react": "*",
+    "@vygruppen/spor-button-react": "*",
+    "@vygruppen/spor-i18n-react": "*",
     "@vygruppen/spor-icon-react": "*"
   },
   "devDependencies": {
