+++ packages/spor-alert-react/src/index.tsx
@@ -1,1 +1,3 @@
-export * from "./Alert";
+export * from "./ClosableAlert";
+export * from "./ExpandableAlert";
+export * from "./StaticAlert";
